requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
flask-login>=0.6.0
flask-jwt-extended>=4.6.0
flask-bcrypt>=1.0.0
flask-mail>=0.10.0
flask-cors>=4.0.0
celery>=5.3.0
redis>=4.6.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
werkzeug>=2.3.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
gunicorn>=21.0.0
pytesseract>=0.3.10
easyocr>=1.7.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0